Issue #ED-2848 merge: 7.0.0 user delete (#9047)
* Issue #ED-2848 feat: added delete user page

* Issue #ED-2848 feat: ui changes for delete button

* Issue #ED-2848 feat: added checkbox amd ui enhancement

* Issue #ED-2848 feat: code changes

* Issue #ED-2848 feat: added text from resourcee files

* Issue #ED-2848 feat: added text from reesourcebundles

* Issue #ED-2848: added labels

* Issue #ED-2848 feat: removed otp-modal

* Issue #ED-2848 feat: code changes

* Issue #ED-2935: Added url for delete user and methods in middleware

* issue #ED-2935 API fix

* issue #ED-2935 API fix

* issue #ED-2935 API fix

* issue #ED-2935 API changes

* issue #ED-2935 API changes

* issue #ED-3123 writing unit testcases for delete user functionality

* issue #ED-3123 writing unit testcases for delete user functionality

* issue #ED-3123 writing unit testcases for delete user functionality

* issue #ED-3123 writing unit testcases for delete user functionality

* issue #ED-3123 writing unit testcases for delete user functionality

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

* Issue #ED-3146 api integration

---------

+++ src/app/client/src/app/plugins/profile/components/delete-account/delete-account.component.ts
@@ -7,6 +7,7 @@ import { Subject } from 'rxjs';
 import { ProfileService } from '../../services';
 import { IInteractEventObject, IInteractEventEdata } from '@sunbird/telemetry';
 import { MatDialog } from '@angular/material/dialog';
+import { DeviceDetectorService } from 'ngx-device-detector';
 
 @Component({
   selector: 'app-delete-account',
@@ -32,7 +33,8 @@ export class DeleteAccountComponent implements OnInit, OnDestroy {
   constructor(public resourceService: ResourceService, public userService: UserService,
     public otpService: OtpService, public toasterService: ToasterService,
     public profileService: ProfileService, private matDialog: MatDialog,
-    public configService: ConfigService,private cacheService:CacheService) { }
+    public configService: ConfigService,private cacheService:CacheService,
+    public deviceDetectorService: DeviceDetectorService) { }
 
   ngOnInit() {
     this.validateAndEditContact();
@@ -60,7 +62,6 @@ export class DeleteAccountComponent implements OnInit, OnDestroy {
           this.resourceService.frmelmnts.lbl.wrongEmailOTP
       };
       this.verifiedUser = false;
-
       this.generateOTP({ request }, otpData);
     }
   }
@@ -113,8 +114,25 @@ export class DeleteAccountComponent implements OnInit, OnDestroy {
   }
 
   verificationSuccess(data) {
-      window.location.replace('/logoff');
-      this.cacheService.removeAll();
+      this.userService.deleteUser().subscribe(
+        (data: ServerResponse) => {
+          if(_.get(data, 'result.response') === 'SUCCESS'){
+            window.location.replace('/logoff');
+            this.cacheService.removeAll();
+            if(this.deviceDetectorService.isMobile()){
+              //TODO changes need to be done on the Mobile Deeplink
+              const url ='dev.sunbird.app://mobile?userId'+ this.userProfile.userId;
+              window.open(url, '_blank');
+            }
+          }
+        },
+        (err) => {
+          //TODO we need to update the error 
+          const errorMessage =  this.resourceService.messages.fmsg.m0085;
+          this.toasterService.error(errorMessage);
+        }
+      );
+       
   }
 
   setInteractEventData() {
@@ -124,7 +142,6 @@ export class DeleteAccountComponent implements OnInit, OnDestroy {
       type: 'click',
       pageid: 'delete-account'
     };
-
     this.telemetryInteractObject = {
       id: this.userService.userid,
       type: 'User',
